# app-engine/src/test/resources/spec/api/openapi_spec.v0.1.0.yml
openapi: 3.0.0
info:
  title: App Engine API
  version: 0.2.0
  description: App Engine API provides a set of requests to interact with tasks and apps.
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8080/api/v1
paths:
  /tasks:
    post:
      summary: Upload and creates a task as a zip archive containing a 'descriptor.yml' file and related task files
      operationId: uploadTask
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                task:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task_description"
        '400':
          description: The descriptor is malformed, e.g. not valid based on the schema, missing files or contains erroneous information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
        '409':
          description: Conflict, a task with the same namespace and version has already been registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'

    get:
      summary: Get descriptions of all tasks available in App Engine
      operationId: getTasks
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/task_description'
  /tasks/{id}:
    get:
      summary: Get task description by UUID
      operationId: getTaskByUUID
      parameters: &task-uuid-only-parameters
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task
      responses: &get-task-description-responses
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task_description'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{namespace}/{version}:
    get:
      summary: Get task description by namespace and version
      operationId: getTaskByNamespaceVersion
      parameters: &task-ns-version-only-parameters
        - in: path
          name: namespace
          required: true
          schema:
            type: string
          description: The namespace of the task
        - in: path
          name: version
          required: true
          schema:
            type: string
          description: The version of the task
      responses: *get-task-description-responses
  /tasks/{id}/inputs:
    get:
      summary: List input parameters by task UUID
      operationId: getTaskInputsByUUID
      parameters: *task-uuid-only-parameters
      responses: &get-task-inputs-responses
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/input_parameter'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{namespace}/{version}/inputs:
    get:
      summary: List input parameters by task namespace and version
      operationId: getTaskInputsByNamespaceVersion
      parameters: *task-ns-version-only-parameters
      responses: *get-task-inputs-responses
  /tasks/{id}/outputs:
    get:
      summary: List output parameters by task UUID
      operationId: getTaskOutputsByUUID
      parameters: *task-uuid-only-parameters
      responses: &get-task-outputs-responses
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/output_parameter'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{namespace}/{version}/outputs:
    get:
      summary: List output parameters by task namespace and version
      operationId: getTaskOutputsByNamespaceVersion
      parameters: *task-ns-version-only-parameters
      responses: *get-task-outputs-responses
  /tasks/{id}/descriptor.yml:
    get:
      summary: Download task descriptor in YAML format
      operationId: getTaskDescriptorByUUID
      parameters: *task-uuid-only-parameters
      responses: &get-task-descriptor-responses
        '200':
          description: Successful response
          content:
            application/yaml:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{namespace}/{version}/descriptor.yml:
    get:
      summary: Download task descriptor in YAML format
      operationId: getTaskDescriptorByNamespaceVersion
      parameters: *task-ns-version-only-parameters
      responses: *get-task-descriptor-responses
...
  /task-runs/{run_id}:
    get:
      summary: Get information about a task run
      operationId: getTaskRun
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task run
      responses:
        '200':
          description: returns the task run description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task_run"
        '404':
          description: task run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    contact:
      type: object
      properties:
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        organization:
          type: string
          description: This person's organization
        email:
          type: string
          description: Email address
        is_contact:
          type: boolean
          description: This person is the contact person
...
    task_description:
      type: object
      required:
        - name
        - namespace
        - version
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Task name
        namespace:
          type: string
          description: Task namespace
        version:
          type: string
          description: Task version
        description:
          type: string
          description: Task description
        authors:
          type: array
          description: List of the task authors
          items:
            $ref: "#/components/schemas/contact"
    enum_item:
      type: string
      maxLength: 256
      minLength: 1
    param-deps-derived-from:
      type: "string"
      pattern: "^inputs/[a-zA-Z0-9_]+$"
    param-deps-matching:
      type: "array"
      items:
        type: string
        pattern: "^(inputs|outputs)/[a-zA-Z0-9_]+$"
    base_parameter:
      type: object
      required:
        - description
        - display_name
        - optional
        - id
        - name
        - type
      properties:
        id:
          type: string
          format: uuid
          description: Parameter App Engine UUID
        name:
          $ref: "#/components/schemas/param_name_type"
        display_name:
          description: Human-readable name for the parameter
          type: "string"
        description:
          description: Human-readable text description of the parameter
          type: string
    type_spec_integer:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          enum: ["integer"]
        gt:
          type: "integer"
          description: 'input constraint to enforce: input > $gt'
        lt:
          type: "integer"
          description: 'input constraint to enforce: input < $lt'
        geq:
          type: "integer"
          description: 'input constraint to enforce: input >= $geq'
        leq:
          type: "integer"
          description: 'input constraint to enforce: input <= $leq'
    type_spec_boolean:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          enum: ["boolean"]
    type_spec_enumeration:
      type: object
      required:
        - id
        - values
      properties:
        id:
          type: string
          enum: ["enumeration"]
        values:
          type: array
          items:
            $ref: "#/components/schemas/enum_item"
    type_spec_string:
      type: object
      required:
        - type
      properties:
        type:
          type: object
          properties:
            id:
              type: string
              enum: ["string"]
            min_length:
              type: integer
              minimum: 0
              description: 'input constraint, string length should be greater or equal to than $min_length'
            max_length:
              type: integer
              minimum: 1
              description: 'input constraint, string length should be less than or equal to $max_length'
    type_spec_number:
      type: object
      required:
        - id
        - infinity_allowed
        - nan_allowed
      properties:
        id:
          type: string
          enum: ["number"]
        gt:
          type: number
          description: 'input constraint to enforce: input > $gt'
        lt:
          type: number
          description: 'input constraint to enforce: input < $lt'
        geq:
          type: number
          description: 'input constraint to enforce: input >= $geq'
        leq:
          type: number
          description: 'input constraint to enforce: input <= $leq'
        nan_allowed:
          type: boolean
          description: 'whether or not nan is accepted as a value for this parameters'
        infinity_allowed:
          type: boolean
          description: 'whether or not infinity is accepted as a value for this parameter'
    type_spec_datetime:
      type: object
      required:
        - type
      properties:
        type:
          type: object
          required:
            - id
          properties:
            id:
              type: string
              enum: ["datetime"]
            before:
              type: string
              format: date-time
              description: the input datetime must be before the specified datetime (inclusive bound). If not specified, no 'before' constraint applied.
            after:
              type: string
              format: date-time
              description: the input datetime must be after the specified datetime (inclusive bound). If not specified, no 'after' constraint applied.
    type_spec_geometry:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          enum: ["geometry"]
    type_spec_image:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          enum: ["image"]
        max_file_size:
          type: integer
          minimum: 1
          description: max image file size, in bytes
        max_width:
          type: integer
          minimum: 1
          description: maximum width for the image
        max_height:
          type: integer
          minimum: 1
          description: maximum height for the image
        formats:
          type: array
          minItems: 1
          items:
            type: string
            enum:
              - jpeg
              - png
              - tiff
    type_spec_array:
      type: object
      required: ["id"]
      properties:
        id:
          type: string
          enum: ["array"]
        subType:
          type: object
          oneOf:
            - $ref: "#/components/schemas/type_spec_integer"
            - $ref: "#/components/schemas/type_spec_boolean"
            - $ref: "#/components/schemas/type_spec_enumeration"
            - $ref: "#/components/schemas/type_spec_string"
            - $ref: "#/components/schemas/type_spec_number"
            - $ref: "#/components/schemas/type_spec_geometry"
            - $ref: "#/components/schemas/type_spec_image"
            - $ref: "#/components/schemas/type_spec_array"
    typed_parameter_integer:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_integer"
    typed_parameter_boolean:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_boolean"
    typed_parameter_enumeration:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_enumeration"
    typed_parameter_string:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_string"
    typed_parameter_number:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_number"
    typed_parameter_datetime:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_datetime"
    typed_parameter_geometry:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_geometry"
    typed_parameter_image:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_image"
    typed_parameter_array:
      allOf:
        - $ref: "#/components/schemas/base_parameter"
        - type: object
          properties:
            dependencies:
              type: object
              properties:
                matching:
                  $ref: "#/components/schemas/param-deps-matching"
        - type: object
          required:
            - type
          properties:
            type:
              $ref: "#/components/schemas/type_spec_array"
    input_parameter:
      type: object
      allOf:
        - oneOf:
          - allOf:
            - $ref: "#/components/schemas/typed_parameter_integer"
            - type: object
              properties:
                default:
                  description: Default integer value. If omitted, no default value
                  type: integer
        - oneOf:
          - allOf:
            - $ref: "#/components/schemas/typed_parameter_boolean"
            - type: object
              properties:
                default:
                  description: Default boolean value. If omitted, no default value
                  type: boolean
        - oneOf:
          - allOf:
            - $ref: "#/components/schemas/typed_parameter_enumeration"
            - type: object
              properties:
                default:
                  description: Default enumeration value. If omitted, no default value
                  type: string
        - oneOf:
          - allOf:
            - $ref: "#/components/schemas/typed_parameter_number"
            - type: object
              properties:
                default:
                  description: Default number value. If omitted, no default value
                  type: number
        - oneOf:
            - allOf:
              - $ref: "#/components/schemas/typed_parameter_string"
              - type: object
                properties:
                  default:
                    description: Default string value. If omitted, no default value
                    type: string
        - oneOf:
            - allOf:
              - $ref: "#/components/schemas/typed_parameter_datetime"
              - type: object
                properties:
                  default:
                    description: Default datetime value. If omitted, no default value
                    type: string
                    format: date-time
        - oneOf:
          - $ref: "#/components/schemas/typed_parameter_geometry"
        - oneOf:
          - $ref: "#/components/schemas/typed_parameter_image"
        - oneOf:
          - $ref: "#/components/schemas/typed_parameter_array"

      properties:
        optional:
          description: Whether or not the parameter is optional
          type: boolean
    output_parameter:
      allOf:
        - oneOf:
          - $ref: "#/components/schemas/typed_parameter_integer"
          - $ref: "#/components/schemas/typed_parameter_boolean"
          - $ref: "#/components/schemas/typed_parameter_enumeration"
          - $ref: "#/components/schemas/typed_parameter_number"
          - $ref: "#/components/schemas/typed_parameter_string"
          - $ref: "#/components/schemas/typed_parameter_datetime"
          - $ref: "#/components/schemas/typed_parameter_geometry"
          - $ref: "#/components/schemas/typed_parameter_image"
          - $ref: "#/components/schemas/typed_parameter_array"
        - type: "object"
          properties:
            dependencies:
              type: "object"
              properties:
                derived_from:
                  $ref: "#/components/schemas/param-deps-derived-from"
    task_run_state:
      type: string
      enum:
        - CREATED      # created but not all inputs provisioned
        - PROVISIONED  # ready to be executed, all inputs have been provisioned
        - QUEUING      # submitting the task to the execution environment
        - QUEUED       # evaluated successfully by execution environment
        - PENDING      # pending execution on the execution environment
        - RUNNING      # task running in the execution environment
        - FAILED       # an error occurred and stopped the process of executing the task (terminal state)
        - FINISHED     # task successfully executed and outputs available (terminal state)
    task_run:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/task_description"
        id:
          type: string
          format: uuid
          description: Parameter App Engine task run UUID
        state:
          $ref: "#/components/schemas/task_run_state"
        created_at:
          description: Creation datetime for the task run
          type: string
          format: date-time
        updated_at:
          description: Last update datetime for the task run
          type: string
          format: date-time
    param_name_type:
      type: string
      pattern: ^[a-zA-Z0-9_]+$
      description: a parameter name
...
    generic_error:
      type: object
      properties:
        errorCode:
          type: string
          pattern: ^APPE-[a-zA-Z0-9_-]+$
          description: an error code structured as `{component}-{reason}`
        message:
          type: string
          description: the human-readable error message associated with the errorCode (non-localized, in english)
        details:
          type: object
          description: provide a more detailed description of the error, structure and content depends on the error code
...